requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
fastapi==0.115.12
    # via -r requirements.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.3.0
    # via
    #   huggingface-hub
    #   torch
google-api-core==2.24.2
    # via google-api-python-client
google-api-python-client==2.165.0
    # via -r requirements.in
google-auth==2.38.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.1
    # via -r requirements.in
googleapis-common-protos==1.69.2
    # via google-api-core
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via langsmith
huggingface-hub==0.29.3
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via torch
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.21
    # via -r requirements.in
langchain-core==0.3.47
    # via
    #   langchain
    #   langchain-text-splitters
langchain-text-splitters==0.3.7
    # via langchain
langsmith==0.3.18
    # via
    #   langchain
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.4
    # via transformers
oauthlib==3.2.2
    # via requests-oauthlib
orjson==3.10.15
    # via langsmith
packaging==24.2
    # via
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   transformers
proto-plus==1.26.1
    # via google-api-core
protobuf==6.30.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.6
    # via
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.27.2
    # via pydantic
pyparsing==3.2.2
    # via httplib2
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-core
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   google-api-core
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   requests-oauthlib
    #   requests-toolbelt
    #   transformers
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9
    # via google-auth
safetensors==0.5.3
    # via transformers
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.39
    # via
    #   -r requirements.in
    #   langchain
starlette==0.46.1
    # via fastapi
sympy==1.13.1
    # via torch
tenacity==9.0.0
    # via langchain-core
tokenizers==0.21.1
    # via transformers
torch==2.6.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.50.0
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r requirements.in
zstandard==0.23.0
    # via langsmith
